MNSIT
MNIST(Modified National Institute of Standards and Technology) 
This is best suited for beginners as it is a real-world dataset where data is already pre-processed, formatted and normalized. Researchers and learners also use it for trying on new algorithms. MNIST is taken as a reference to develop other such datasets.